API cleanup
