--- Editor/CustomEditors/InspectorReadOnlyPropertyDrawer.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: aba5af1c1467469b8bdc3714bc85ce36
-timeCreated: 1657775439
