+++ Runtime/Authoring/Conversion/ClipEventsBlobConverter.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 6c459b9b42064336b2fbab6190c2b0e2
+timeCreated: 1658438198
