--- Runtime/Authoring/InspectorReadOnly.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: a3f058a0e4344f39be09df118e91ef06
-timeCreated: 1657775377
