+++ Runtime/Authoring/RootMotionMode.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b6d9f9957b244df98c0ddc327cd16249
+timeCreated: 1658438379
