+++ Runtime/Components/AnimationParameters.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: bfbdf935582440ddbbf94cdbb56f74fd
+timeCreated: 1658438842
