+++ Runtime/Systems/UpdateStateMachineJob.cs
@@ -3,7 +3,6 @@ using Latios.Kinemation;
 using Unity.Burst;
 using Unity.Entities;
 using Unity.Mathematics;
-using UnityEngine;
 
 namespace DOTSAnimation
 {
