--- Runtime/Utils/AnimationUtils.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: e90bc1379c054df4a13cc5b809c064e3
-timeCreated: 1655354193
