--- Runtime/Utils/CollectionUtils.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 4a449782eaac4c11b6d6cdddc8090897
-timeCreated: 1655664904
