--- Runtime/Utils/LinearBlendSampling.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: dccf5178a5024cfa808fe5bd160fc656
-timeCreated: 1655788579
