--- Runtime/Utils/StateMachineUtils.cs
@@ -1,12 +1,0 @@
-﻿using Latios.Kinemation;
-using Unity.Entities;
-using Unity.Mathematics;
-using UnityEngine;
-using UnityEngine.Assertions;
-
-namespace DOTSAnimation
-{
-    internal static class StateMachineUtils
-    {
-    }
-}
