--- Runtime/Utils/StateMachineUtils.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 9d8927357ad44f009eefa50287ea8194
-timeCreated: 1655905178
